Study-01/web_version/api/requirements.txt
Pillow==10.0.0
numpy==1.24.3
tensorflow==2.13.0
